Added comments and moved proposal tx events logic to hook.
