+++ packages/stateful/hooks/index.ts
@@ -3,6 +3,7 @@ export * from './contracts'
 export * from './useAwaitNextBlock'
 export * from './useDaoProposalSinglePublishProposal'
 export * from './useEncodedDaoProposalSinglePrefill'
+export * from './useExecutedProposalTxEvents'
 export * from './useInbox'
 export * from './useLoadingDaoCardInfos'
 export * from './usePinnedDaos'
